Factor out frontend/graphql.py.
