+++ frontend/tests/util.py
@@ -1,4 +1,4 @@
-from frontend.views import FRONTEND_QUERY_DIR, add_graphql_fragments
+from frontend.graphql import FRONTEND_QUERY_DIR, add_graphql_fragments
 
 
 def get_frontend_query(filename) -> str:
